+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.5.0 : TBD
+## 0.5.1 : TBD
+
+- **Added**: prefix to tailwind
+- **Changed**: edit `save layer` functions
+
+## 0.5.0 : 2024-07-16
 
 - **Added**: Edit mode
 - **Added**: New params to viewer